Re-namespace user-related controllers
