+++ src/Controller/User/UserApiKeyController.php
@@ -2,34 +2,37 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\User;
 
-use App\Response\EmptyBody;
 use App\Response\ErrorResponse;
 use App\Response\ErrorResponseBody;
+use App\Response\LabelledBody;
+use App\Response\LabelledCollectionBody;
 use App\Response\Response;
+use App\Response\User\ApiKey;
 use App\Security\AuthenticationToken;
-use App\Security\RefreshToken;
-use App\Security\UserId;
 use Psr\Http\Client\ClientExceptionInterface;
+use SmartAssert\ServiceClient\Exception\InvalidResponseDataException;
+use SmartAssert\ServiceClient\Exception\InvalidResponseTypeException;
 use SmartAssert\ServiceClient\Exception\NonSuccessResponseException;
 use SmartAssert\UsersClient\Client;
 use SmartAssert\UsersClient\Exception\UnauthorizedException;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
 
-readonly class RefreshTokenController
+#[Route('/user/apikey', name: 'user_apikey_')]
+readonly class UserApiKeyController
 {
     public function __construct(
         private Client $client
     ) {
     }
 
-    #[Route('/user/refresh_token/revoke-all', name: 'user_revoke_all_refresh_token', methods: ['POST'])]
-    public function revokeAllForUser(AuthenticationToken $token, UserId $userId): JsonResponse
+    #[Route('/list', name: 'list', methods: ['GET'])]
+    public function list(AuthenticationToken $token): JsonResponse
     {
         try {
-            $this->client->revokeFrontendRefreshTokensForUser($token->token, $userId->id);
+            $apiKeyCollection = $this->client->listUserApiKeys($token->token);
         } catch (ClientExceptionInterface $e) {
             $code = $e->getCode();
             $message = $e->getMessage();
@@ -42,6 +45,33 @@ readonly class RefreshTokenController
                         'error' => [
                             'code' => $code,
                             'message' => $message,
+                        ],
+                    ]
+                )
+            );
+        } catch (InvalidResponseDataException $e) {
+            return new ErrorResponse(
+                new ErrorResponseBody(
+                    'invalid-response-data',
+                    [
+                        'service' => 'users',
+                        'data' => $e->getResponse()->getBody(),
+                        'data-type' => [
+                            'expected' => $e->expected,
+                            'actual' => $e->actual,
+                        ],
+                    ]
+                )
+            );
+        } catch (InvalidResponseTypeException $e) {
+            return new ErrorResponse(
+                new ErrorResponseBody(
+                    'invalid-response-type',
+                    [
+                        'service' => 'users',
+                        'content-type' => [
+                            'expected' => $e->expected,
+                            'actual' => $e->actual,
                         ],
                     ]
                 )
@@ -68,16 +98,21 @@ readonly class RefreshTokenController
             );
         }
 
+        $apiKeys = [];
+        foreach ($apiKeyCollection as $usersClientApiKey) {
+            $apiKeys[] = new ApiKey($usersClientApiKey->label, $usersClientApiKey->key);
+        }
+
         return new Response(
-            new EmptyBody()
+            new LabelledCollectionBody('api_keys', $apiKeys)
         );
     }
 
-    #[Route('/user/refresh_token/revoke', name: 'user_revoke_refresh_token', methods: ['POST'])]
-    public function revoke(AuthenticationToken $token, RefreshToken $refreshToken): JsonResponse
+    #[Route('/', name: 'get_default', methods: ['GET'])]
+    public function getDefault(AuthenticationToken $token): JsonResponse
     {
         try {
-            $this->client->revokeFrontendRefreshToken($token->token, $refreshToken->refreshToken);
+            $apiKey = $this->client->getUserDefaultApiKey($token->token);
         } catch (ClientExceptionInterface $e) {
             $code = $e->getCode();
             $message = $e->getMessage();
@@ -90,6 +125,33 @@ readonly class RefreshTokenController
                         'error' => [
                             'code' => $code,
                             'message' => $message,
+                        ],
+                    ]
+                )
+            );
+        } catch (InvalidResponseDataException $e) {
+            return new ErrorResponse(
+                new ErrorResponseBody(
+                    'invalid-response-data',
+                    [
+                        'service' => 'users',
+                        'data' => $e->getResponse()->getBody(),
+                        'data-type' => [
+                            'expected' => $e->expected,
+                            'actual' => $e->actual,
+                        ],
+                    ]
+                )
+            );
+        } catch (InvalidResponseTypeException $e) {
+            return new ErrorResponse(
+                new ErrorResponseBody(
+                    'invalid-response-type',
+                    [
+                        'service' => 'users',
+                        'content-type' => [
+                            'expected' => $e->expected,
+                            'actual' => $e->actual,
                         ],
                     ]
                 )
@@ -116,8 +178,15 @@ readonly class RefreshTokenController
             );
         }
 
+        if (null === $apiKey) {
+            return new JsonResponse(null, 404);
+        }
+
         return new Response(
-            new EmptyBody()
+            new LabelledBody(
+                'api_key',
+                new ApiKey($apiKey->label, $apiKey->key)
+            )
         );
     }
 }
